refactor(core-weave): extract slice classification helpers
- move pending weave slice detection into slice_classification.ts
- move shared artifact-history query helpers into artifact_history_queries.ts
- preserve detectPendingWeaveSlice through the weave.ts compatibility façade
- update the codebase overview with the extracted helper modules
- verify with fmt, lint, check, core weave tests, integration weave tests, and import-cycle audit
